requirements.txt
einops~=0.3.0
numpy~=1.19.5
pandas~=1.2.4
pytorch-lightning
pyyaml~=5.3.1
scikit-learn~=0.24.2
scipy~=1.6.3
tables
tensorboard
torch~=1.8.1
torchmetrics~=0.5.1
tensorflow~=2.5.0
tensorflow-gpu~=2.4.0
fancyimpute~=0.6.0